Split plotting into two classes
